% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissing.R
\name{checkM}
\alias{checkM}
\title{Check number of missing values}
\usage{
checkM(vect, nonM = FALSE)
}
\arguments{
\item{vect}{A numeric vector of values.}

\item{nonM}{A logical indicating whether to report the non-missing or missing counts.}
}
\description{
This is a function to count the number of missing values.
}
\examples{
x <- c(1:100, NA)
y <- checkM(x)
}
