% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpperLower.R
\name{get_tri}
\alias{get_tri}
\title{Get Lower or Upper Triangles}
\usage{
get_tri(cormat, upper = FALSE)
}
\arguments{
\item{cormat}{a matrix.}

\item{upper}{a logical.}
}
\description{
This is a function to obtain the lower or upper triangle of a matrix.
}
