% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inver_norm.R
\name{inver_norm}
\alias{inver_norm}
\title{Transform variables to normal}
\usage{
inver_norm(x, keep_mean_sd = TRUE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{keep_mean_sd}{a logical indicating whether mean and sd should be kept.}
}
\description{
This is a function to apply rank-based transformation to a variable  to have a normal distribution with
mean 0 and variance 1. Missing values in  are acceptable.
}
\examples{
x <- stats::rchisq( n = 100, df=4)
y <- inver_norm(x)
}
