% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorizef.R
\name{winsorizef}
\alias{winsorizef}
\title{Winsorize variables to eliminate outliers}
\usage{
winsorizef(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\description{
This is a function to apply winsorize a variable to
remove large values outside of the 3.5 standard deviations (usually for normal variables).
}
\examples{
x <- stats::rchisq( n = 100, df=4)
y <- winsorizef(x)
}
